+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ## Unreleased
 
-- Fix regression where mail search would fail on non-ascii search criteria (#10121)
 - Fix so distribution packages (and composer.json) don't include development dependencies
+- Fix regression where mail search would fail on non-ascii search criteria (#10121)
+- Fix regression where some data url images could get ignored/lost (#10128)
 
 ## Release 1.5.14